+++ src/main/java/com/m/roman/oreh/repository/MenuDeliveryRepository.java
@@ -3,5 +3,5 @@ package com.m.roman.oreh.repository;
 import com.m.roman.oreh.entity.MenuDelivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuDeliveryRepository extends JpaRepository<MenuDelivery, Integer> {
+public interface MenuDeliveryRepository extends JpaRepository<MenuDelivery, Long> {
 }
