+++ src/main/java/com/m/roman/oreh/repository/ProductRepository.java
@@ -7,8 +7,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     @Query("select p from Product p")
     Slice<Product> getAllProducts(int page, int size);
+
+
 }
