+++ src/main/java/com/m/roman/oreh/repository/SpecificationRepository.java
@@ -3,5 +3,5 @@ package com.m.roman.oreh.repository;
 import com.m.roman.oreh.entity.Specification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpecificationRepository extends JpaRepository<Specification, Integer> {
+public interface SpecificationRepository extends JpaRepository<Specification, Long> {
 }
