+++ src/main/java/com/m/roman/oreh/repository/TitleTextRepository.java
@@ -3,5 +3,5 @@ package com.m.roman.oreh.repository;
 import com.m.roman.oreh.entity.TitleText;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TitleTextRepository extends JpaRepository<TitleText, Integer> {
+public interface TitleTextRepository extends JpaRepository<TitleText, Long> {
 }
