+++ src/main/java/com/m/roman/oreh/service/MediaService.java
@@ -9,11 +9,11 @@ public interface MediaService {
 
     public List<Media> getAllMedia();
 
-    public Media getMedia(int id);
+    public Media getMedia(long id);
 
     public void saveMedia(Media media);
 
-    public void deleteMedia(int id);
+    public void deleteMedia(long id);
 
     public void updateMedia(Media media);
