+++ src/main/java/com/m/roman/oreh/service/NewsService.java
@@ -8,11 +8,11 @@ public interface NewsService {
 
     public List<News> getAllNews();
 
-    public News getNews(int id);
+    public News getNews(long id);
 
     public void saveNews(News news);
 
-    public void deleteNews(int id);
+    public void deleteNews(long id);
 
     public void updateNews(News news);
 }
