+++ src/main/java/com/m/roman/oreh/service/ProductService.java
@@ -6,7 +6,6 @@ import org.springframework.data.domain.Slice;
 import java.util.List;
 
 public interface ProductService {
-    // Пагинация на уровне интерфейса сервиса помощью Slice
 
     public Slice<Product> getAllProducts(int page, int size);
 
@@ -14,10 +13,10 @@ public interface ProductService {
 
     public void saveProduct(Product product);
 
-    public void deleteProduct(int id);
+    public void deleteProduct(long id);
 
     public void updateProduct(Product product);
 
-    public Product getProduct(int id);
+    public Product getProduct(long id);
 }
