+++ src/main/java/com/m/roman/oreh/service/TitleTextService.java
@@ -8,11 +8,11 @@ public interface TitleTextService {
 
     public void saveTitleText(TitleText titleText);
 
-    public void deleteTitleText(int id);
+    public void deleteTitleText(long id);
 
     public void updateTitleText(TitleText titleText);
 
     public List<TitleText> getAllTitleText();
 
-    public TitleText getTitleText(int id);
+    public TitleText getTitleText(long id);
 }
